pyproject.toml
[project]
name = "quent"
version = "0.1.0"
description = "Moving-average trend-following backtest and trading system."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2",
    "numpy>=2.0",
    "PyYAML>=6.0",
    "matplotlib>=3.9",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py313"

